Refactor for consistency
This puts the different code katas into their own folders.
